Initial commit / upload
Submitting my changes from an older branch of an older repo that really
should have been its own repo to begin with.

The bot’s auth key shall always remain out of this repo for security
and is kept in a separate .json file called token.json and the path is
relative to where the bot runs in an alpine container. The docker file
and security files will also stay out of this repo.